Move attendance viewsets and serializers to viewsets/ package
